Updated README and some more code refactoring to run code for minmal use

+++ pkg/service/v1/feedback-service.go
@@ -21,7 +21,6 @@ const (
 )
 
 type feedbackServiceServer struct {
-	client *mongo.Client
 	db     *mongo.Database
 	logger *logrus.Logger
 }
@@ -71,8 +70,8 @@ var can = "can"
 var topicsCovered = "We have covered following topics in '%v':"
 
 // NewFeedbackServiceServer creates FeedbackServiceServer
-func NewFeedbackServiceServer(client *mongo.Client, db *mongo.Database, logger *logrus.Logger) v1.FeedbackServiceServer {
-	return &feedbackServiceServer{client: client, db: db, logger: logger}
+func NewFeedbackServiceServer(db *mongo.Database, logger *logrus.Logger) v1.FeedbackServiceServer {
+	return &feedbackServiceServer{db: db, logger: logger}
 }
 
 func (fs *feedbackServiceServer) checkAPI(api string) error {
@@ -86,41 +85,42 @@ func (fs *feedbackServiceServer) checkAPI(api string) error {
 }
 
 func (fs *feedbackServiceServer) Create(ctx context.Context, req *v1.FeedbackRequest) (*v1.FeedbackResponse, error) {
-
 	if err := fs.checkAPI(req.Api); err != nil {
 		return new(v1.FeedbackResponse), err
 	}
 	fReq := req.GetFeedbackReq()
-	// Create new request Id
-	fReq.Id = primitive.NewObjectID().Hex()
-
-	coll := fs.db.Collection("feedback_request")
-	//Insert created FeedbackRequest
-	result, err := coll.InsertOne(ctx, fReq)
-
-	if result.InsertedID == nil && err != nil {
-		fs.logger.WithFields(logrus.Fields{
-			"request": fReq,
-			"status":  http.StatusInternalServerError,
-			"Error":   err,
-		}).Error("Unable Insert the Document!")
-
-		fRes := &v1.FeedbackResponse{
-			Api:         "v1",
-			StatusCode:  http.StatusInternalServerError,
-			RequestId:   "",
-			Message:     fmt.Sprintln(err),
-			FeedbackRes: nil,
-		}
-		return fRes, nil
-	}
-
-	fResult := coll.FindOne(ctx, bson.D{primitive.E{Key: "id", Value: fReq.Id}})
 	feedbackRes := v1.Feedback{}
 
-	if err := fResult.Decode(&feedbackRes); err != nil {
-		logrus.Errorf("Unable to read document for request id: %v", fReq.Id)
-		return nil, nil
+	if fs.db != nil {
+		// Create new request Id
+		fReq.Id = primitive.NewObjectID().Hex()
+
+		coll := fs.db.Collection("feedback_request")
+		//Insert created FeedbackRequest
+		result, err := coll.InsertOne(ctx, fReq)
+
+		if result.InsertedID == nil && err != nil {
+			fs.logger.WithFields(logrus.Fields{
+				"request": fReq,
+				"status":  http.StatusInternalServerError,
+				"Error":   err,
+			}).Error("Unable Insert the Document!")
+
+			fRes := &v1.FeedbackResponse{
+				Api:         "v1",
+				StatusCode:  http.StatusInternalServerError,
+				RequestId:   "",
+				Message:     fmt.Sprintln(err),
+				FeedbackRes: nil,
+			}
+			return fRes, nil
+		}
+
+		fResult := coll.FindOne(ctx, bson.D{primitive.E{Key: "id", Value: fReq.Id}})
+		if err := fResult.Decode(&feedbackRes); err != nil {
+			logrus.Errorf("Unable to read document for request id: %v", fReq.Id)
+			return nil, nil
+		}
 	}
 
 	fRes := &v1.FeedbackResponse{
@@ -134,18 +134,20 @@ func (fs *feedbackServiceServer) Create(ctx context.Context, req *v1.FeedbackReq
 }
 
 func (fs *feedbackServiceServer) Read(ctx context.Context, req *v1.ReadFeedbackRequest) (*v1.FeedbackResponse, error) {
-	coll := fs.db.Collection("feedback_request")
-	fResult := coll.FindOne(ctx, bson.D{primitive.E{Key: "id", Value: req.RequestId}})
+
 	feedbackRes := v1.Feedback{}
 	fRes := &v1.FeedbackResponse{}
 	fRes.Api = "v1"
 	fRes.RequestId = req.RequestId
-
-	if err := fResult.Decode(&feedbackRes); err != nil {
-		logrus.Errorf("Unable to read document for request id: %v", req.RequestId)
-		fRes.StatusCode = http.StatusNotFound
-		fRes.Message = fmt.Sprint(fResult.Decode(bson.M{}))
-		return fRes, nil
+	if fs.db != nil {
+		coll := fs.db.Collection("feedback_request")
+		fResult := coll.FindOne(ctx, bson.D{primitive.E{Key: "id", Value: req.RequestId}})
+		if err := fResult.Decode(&feedbackRes); err != nil {
+			logrus.Errorf("Unable to read document for request id: %v", req.RequestId)
+			fRes.StatusCode = http.StatusNotFound
+			fRes.Message = fmt.Sprint(fResult.Decode(bson.M{}))
+			return fRes, nil
+		}
 	}
 
 	fRes.StatusCode = http.StatusCreated
@@ -155,37 +157,45 @@ func (fs *feedbackServiceServer) Read(ctx context.Context, req *v1.ReadFeedbackR
 	return fRes, nil
 }
 func (fs *feedbackServiceServer) GenerateFeedbackForRequest(ctx context.Context, req *v1.GenerateFeedbackRequest) (*v1.GeneratedFeedbackResponse, error) {
-	coll := fs.db.Collection("feedback_request")
-	fResult := coll.FindOne(ctx, bson.D{primitive.E{Key: "id", Value: req.FeedbackReq.Id}})
 	gfRes := &v1.GeneratedFeedbackResponse{}
 	gfRes.Api = "v1"
 
 	fRes := v1.Feedback{}
-	if err := fResult.Decode(&fRes); err != nil {
-		logrus.Errorf("Unable to read document for request id: %v", req.FeedbackReq.Id)
-	}
-	req.FeedbackReq = &fRes
-	gfRes, err := generateFeedbackReport(req)
+	if fs.db != nil {
+		coll := fs.db.Collection("feedback_request")
+		fResult := coll.FindOne(ctx, bson.D{primitive.E{Key: "id", Value: req.FeedbackReq.Id}})
 
-	return gfRes, err
+		if err := fResult.Decode(&fRes); err != nil {
+			logrus.Errorf("Unable to read document for request id: %v", req.FeedbackReq.Id)
+		}
+		req.FeedbackReq = &fRes
+		gfRes, err := generateFeedbackReport(req)
+		return gfRes, err
+	}
+
+	return gfRes, nil
 }
 
 func (fs *feedbackServiceServer) Delete(ctx context.Context, req *v1.DeleteFeedbackRequest) (*v1.DeleteFeedbackResponse, error) {
-	coll := fs.db.Collection("feedback_request")
-	result, err := coll.DeleteOne(ctx, bson.D{primitive.E{Key: "id", Value: req.RequestId}})
-
-	if err != nil {
-		fs.logger.Errorf("Unable to delete document for request id: %v Deleted Count: %v", req.RequestId, result.DeletedCount)
-		return nil, nil
-	}
 	var fRes = &v1.DeleteFeedbackResponse{}
 	fRes.Api = "v1"
-	if result.DeletedCount == 0 {
-		fRes.StatusCode = http.StatusNotFound
-	} else {
-		fRes.StatusCode = http.StatusOK
-	}
 
+	if fs.db != nil {
+		coll := fs.db.Collection("feedback_request")
+		result, err := coll.DeleteOne(ctx, bson.D{primitive.E{Key: "id", Value: req.RequestId}})
+
+		if err != nil {
+			fs.logger.Errorf("Unable to delete document for request id: %v Deleted Count: %v", req.RequestId, result.DeletedCount)
+			return nil, nil
+		}
+		if result.DeletedCount == 0 {
+			fRes.StatusCode = http.StatusNotFound
+		} else {
+			fRes.StatusCode = http.StatusOK
+		}
+		return fRes, nil
+	}
+	fRes.StatusCode = http.StatusOK
 	return fRes, nil
 }
 func (fs *feedbackServiceServer) GenerateFeedbackFromFormData(ctx context.Context, req *v1.GenerateFeedbackRequest) (*v1.GeneratedFeedbackResponse, error) {
